// human-resource-management-app/src/models/employee.ts
export default interface Employee {
    maNhanVien: string;
    hoTen: string;
    ngaySinh: Date;
    gioiTinh: string;
    cccd: string;
    tonGiao: string;
    diaChi: string;
    soDienThoai: string;
    email: string;
    ngayVaoLam: Date;
    mucLuong: number;
    trangThai: string;
    phongBanId: string;
    chucVuId: string;
    soNguoiPhuThuoc: string;
    trangThaiHonNhan: string;
}